requirements.txt
python-telegram-bot==21.6
nest_asyncio
google-generativeai
python-dotenv
